did not plan on actually implementing tr_polys with forward+ (it didn't even solve the issue of incadescent light panels not working right because they're baked into worldspawn)

+++ engine/inc/uf/engine/graph/graph.h
@@ -72,7 +72,7 @@ namespace pod {
 				uf::stl::string tag = "worldspawn";
 				uf::stl::string player = "info_player_spawn";
 				
-				uint64_t hash = 0;
+				size_t hash = 0;
 				float lastUpdate = 0;
 			} stream;
 		} settings;
